chore: StepBinder 파일 분리
